Edicion del open space
